Make fish_indent available as a builtin
